pyproject.toml
[tool.poetry]
name = "jatszohaz-algo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
numpy = "^2.1.1"
gspread = "^6.1.2"
google-auth = "^2.35.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
watchdog = "^5.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
